+++ packages/util-package-extender/HISTORY.md
@@ -1,5 +1,12 @@
 # History
 
+## 0.0.4 (2019-09-25)
+    * Convert `getPackageExtensionDetails` to `async/await`
+	* Improve `getPackageExtensionDetails`
+		* Checks for correct formatting of the `extendsPackage` key
+		* Checks for correct semver versioning in the `extendsPackage` key
+		* Checks `extendsPackage` exists on NPM
+
 ## 0.0.3 (2019-09-25)
     * Standardised code locations across monorepo